wire the authz client
